Extract BoolStruct and BoolStructTree
Extract BoolStruct and BoolStructTree into separate modules
