+++ lib/haskell/anyall/src/AnyAll.hs
@@ -3,10 +3,12 @@ module AnyAll
     , module AnyAll.SVGLadder
     , module AnyAll.Relevance
     , module AnyAll.PP
+    , module AnyAll.BoolStruct
     ) where
 
 import AnyAll.Types
 import AnyAll.SVGLadder
 import AnyAll.Relevance
 import AnyAll.PP
+import AnyAll.BoolStruct
