+++ lib/haskell/anyall/test/AnyAll/TypesSpec.hs
@@ -6,76 +6,14 @@ import Data.Text (Text)
 import Data.Aeson (decode, encode)
 import qualified Data.Map as Map
 import AnyAll.Types
-import Prelude hiding (any, all)
-import AnyAll.Types (Formula(FAtom))
-import Data.Tree ( Tree( Node ))
 
 type MarkingMap = Map.Map Text (Default Bool)
 
 markingMap :: Either (Maybe Bool) (Maybe Bool) -> Map.Map Text (Default Bool)
 markingMap payload = Map.singleton "key" (Default payload)
 
-all :: [BoolStruct (Maybe l) a] -> BoolStruct (Maybe l) a
-all = All Nothing
-
-allPre :: Text -> [BoolStruct (Maybe (Label Text)) a] -> BoolStruct (Maybe (Label Text)) a
-allPre label = All (Just (Pre label))
-
-anyPre :: Text -> [BoolStruct (Maybe (Label Text)) a] -> BoolStruct (Maybe (Label Text)) a
-anyPre label = All (Just (Pre label))
-
-any :: [BoolStruct (Maybe l) a ] -> BoolStruct (Maybe l) a
-any = Any Nothing
-
-atom :: Text -> BoolStruct (Maybe Text) Text
-atom = Leaf
-
-atomNode :: Text -> Tree (Maybe Text, Formula Text)
-atomNode t = Node (Nothing :: Maybe Text, FAtom t) []
-
-allDt ::  [Tree (Maybe Text, Formula Text)] -> Tree (Maybe Text, Formula Text)
-allDt = Node (Nothing :: Maybe Text, FAll :: Formula Text)
-
-anyDt ::  [Tree (Maybe Text, Formula Text)] -> Tree (Maybe Text, Formula Text)
-anyDt = Node (Nothing :: Maybe Text, FAny :: Formula Text)
-
-type WireBoolStruct = BoolStruct (Maybe (Label Text)) Text
-
-leaf :: Text -> WireBoolStruct
-leaf = Leaf
-
 spec :: Spec
 spec = do
-  describe "simplifyItem" $ do
-    it "should leave Leaf " $ do
-      simplifyItem (leaf "foo") `shouldBe` leaf "foo"
-
-    it "should elevate nested All children" $ do
-      simplifyItem ( all
-                      [ all [leaf "foo1", leaf "foo2"],
-                        all [leaf "bar", all [leaf "bat"]],
-                        allPre "something"
-                          [ leaf "baz",
-                            allPre "something" [leaf "bbb"]
-                          ],
-                        anyPre "something" [leaf "qux"]
-                      ] )
-        `shouldBe` all
-                    [ leaf "foo1",
-                      leaf "foo2",
-                      leaf "bar",
-                      leaf "bat",
-                      allPre "something" [leaf "baz", leaf "bbb"],
-                      leaf "qux"
-                    ]
-
-    it "should simplify singleton Leaf" $ do
-      simplifyItem ( all [leaf "foo"] )
-        `shouldBe`   leaf "foo"
-    it "should simplify not-nots" $ do
-      simplifyItem ( Not $ Not $ leaf "not" )
-        `shouldBe`   leaf "not"
-
   describe "labelFirst" $ do
     it "extracts the only from Pre" $ do
       labelFirst (Pre "a") `shouldBe` "a"
@@ -157,43 +95,3 @@ spec = do
       let ma = markingMap $ Right $ Just False
           q = encode Marking {getMarking = ma}
       q `shouldBe` "{\"getMarking\":{\"key\":{\"getDefault\":{\"Right\":false}}}}"
-
-  describe "nnf transformation" $ do
-    let
-        a = atom "a"
-        b = atom "b"
-
-    it "nnf (not (not a)) == a" $ do
-      nnf (Not . Not $ a) `shouldBe` a
-
-    it "nnf (not (not all [not not a, not not b])) == all [a, b]" $ do
-      nnf (Not . Not $ all [Not . Not $ a, Not . Not $ b]) `shouldBe` all [a, b]
-
-    it "nnf (all [not not a, not not b]) == (all [a,b])" $ do
-      nnf (all [Not . Not $ a, Not . Not $ b]) `shouldBe` all [a, b]
-
-    it "nnf (not (all [a, b])) == (any [not a, not b])" $ do
-      nnf (Not $ all [a, b]) `shouldBe` any [Not a, Not b]
-
-    it "nnf (not (any a)) == (all (not b))" $ do
-      nnf (Not $ any [a, b]) `shouldBe` all [ Not a, Not b]
-
-  describe "nnfdt transformation" $ do
-    let
-        at = atomNode "a"
-        bt = atomNode "b"
-
-    it "nnf (not (not a)) == a" $ do
-      nnfDT (notDt . notDt $ at) `shouldBe` at
-
-    it "nnf (not (all [a, b])) == (any [not a, not b])" $ do
-      nnfDT (notDt $ allDt [at, bt]) `shouldBe` anyDt [notDt at, notDt bt]
-
-    it "nnf (not (any [a, b])) == (all [not a, not b])" $ do
-      nnfDT (notDt $ anyDt [at, bt]) `shouldBe` allDt [notDt at, notDt bt]
-
-    it "nnf (not (not all [not not a, not not b])) == all [a, b]" $ do
-      nnfDT (notDt . notDt $ allDt [notDt . notDt $ at, notDt . notDt $ bt]) `shouldBe` allDt [at, bt]
-
-    it "nnf (all [not not a, not not b]) == (all [a,b])" $ do
-      nnfDT (allDt [notDt . notDt $ at, notDt . notDt $ bt]) `shouldBe` allDt [at, bt]
